project1-be/pyproject.toml
[tool.poetry]
name = "project1-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
celery-stubs = "^0.1.3"
celery-types = "^0.22.0"
django-types = "^0.19.1"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.4"
django-extensions = "^3.2.3"
watchdog = "^4.0.1"  # technically a dev dependency, but it's in the docker container, so it is run in prod to try and keep the container the same in dev as in prod. Used for restarting Celery workers
invoke = "^2.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
djangorestframework-stubs = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
